Make new

+++ include/sass/functions.h
@@ -51,9 +51,9 @@ enum Sass_Callee_Type {
 };
 
 // Creator for sass custom importer return argument list
-ADDAPI Sass_Importer_List ADDCALL sass_make_importer_list (size_t length);
-ADDAPI Sass_Importer_Entry ADDCALL sass_importer_get_list_entry (Sass_Importer_List list, size_t idx);
-ADDAPI void ADDCALL sass_importer_set_list_entry (Sass_Importer_List list, size_t idx, Sass_Importer_Entry entry);
+ADDAPI Sass_Importer_List ADDCALL sass_make_importer_list (uint32_t length);
+ADDAPI Sass_Importer_Entry ADDCALL sass_importer_get_list_entry (Sass_Importer_List list, uint32_t idx);
+ADDAPI void ADDCALL sass_importer_set_list_entry (Sass_Importer_List list, uint32_t idx, Sass_Importer_Entry entry);
 ADDAPI void ADDCALL sass_delete_importer_list (Sass_Importer_List list);
 
 
@@ -70,24 +70,24 @@ ADDAPI void* ADDCALL sass_importer_get_cookie (Sass_Importer_Entry cb);
 ADDAPI void ADDCALL sass_delete_importer (Sass_Importer_Entry cb);
 
 // Creator for sass custom importer return argument list
-ADDAPI Sass_Import_List ADDCALL sass_make_import_list (size_t length);
+ADDAPI Sass_Import_List ADDCALL sass_make_import_list (uint32_t length);
 // Creator for a single import entry returned by the custom importer inside the list
 ADDAPI Sass_Import_Entry ADDCALL sass_make_import_entry (const char* path, char* source, char* srcmap);
 ADDAPI Sass_Import_Entry ADDCALL sass_make_import(const char* imp_path, const char* abs_base, char* source, char* srcmap);
 ADDAPI Sass_Import_Entry ADDCALL sass_make_import2(const char* imp_path, const char* abs_base, char* source, char* srcmap, enum Sass_Import_Type type);
 // set error message to abort import and to print out a message (path from existing object is used in output)
-ADDAPI Sass_Import_Entry ADDCALL sass_import_set_error(Sass_Import_Entry import, const char* message, size_t line, size_t col);
+ADDAPI Sass_Import_Entry ADDCALL sass_import_set_error(Sass_Import_Entry import, const char* message, uint32_t line, uint32_t col);
 
 // Setters to insert an entry into the import list (you may also use [] access directly)
 // Since we are dealing with pointers they should have a guaranteed and fixed size
-ADDAPI void ADDCALL sass_import_set_list_entry (Sass_Import_List list, size_t idx, Sass_Import_Entry entry);
-ADDAPI Sass_Import_Entry ADDCALL sass_import_get_list_entry (Sass_Import_List list, size_t idx);
+ADDAPI void ADDCALL sass_import_set_list_entry (Sass_Import_List list, uint32_t idx, Sass_Import_Entry entry);
+ADDAPI Sass_Import_Entry ADDCALL sass_import_get_list_entry (Sass_Import_List list, uint32_t idx);
 
 // Getters for callee entry
 ADDAPI const char* ADDCALL sass_callee_get_name (Sass_Callee_Entry);
 ADDAPI const char* ADDCALL sass_callee_get_path (Sass_Callee_Entry);
-ADDAPI size_t ADDCALL sass_callee_get_line (Sass_Callee_Entry);
-ADDAPI size_t ADDCALL sass_callee_get_column (Sass_Callee_Entry);
+ADDAPI uint32_t ADDCALL sass_callee_get_line (Sass_Callee_Entry);
+ADDAPI uint32_t ADDCALL sass_callee_get_column (Sass_Callee_Entry);
 ADDAPI enum Sass_Callee_Type ADDCALL sass_callee_get_type (Sass_Callee_Entry);
 
 // Getters and Setters for environments (lexical, local and global)
@@ -110,8 +110,8 @@ ADDAPI enum Sass_Import_Type ADDCALL sass_import_get_type(Sass_Import_Entry);
 //ADDAPI char* ADDCALL sass_import_take_source (Sass_Import_Entry);
 //ADDAPI char* ADDCALL sass_import_take_srcmap (Sass_Import_Entry);
 // Getters from import error entry
-ADDAPI size_t ADDCALL sass_import_get_error_line (Sass_Import_Entry);
-ADDAPI size_t ADDCALL sass_import_get_error_column (Sass_Import_Entry);
+ADDAPI uint32_t ADDCALL sass_import_get_error_line (Sass_Import_Entry);
+ADDAPI uint32_t ADDCALL sass_import_get_error_column (Sass_Import_Entry);
 ADDAPI const char* ADDCALL sass_import_get_error_message (Sass_Import_Entry);
 
 // Deallocator for associated memory (incl. entries)
@@ -122,14 +122,14 @@ ADDAPI void ADDCALL sass_delete_import (Sass_Import_Entry);
 
 
 // Creators for sass function list and function descriptors
-ADDAPI Sass_Function_List ADDCALL sass_make_function_list (size_t length);
+ADDAPI Sass_Function_List ADDCALL sass_make_function_list (uint32_t length);
 ADDAPI Sass_Function_Entry ADDCALL sass_make_function (const char* signature, Sass_Function_Fn cb, void* cookie);
 ADDAPI void ADDCALL sass_delete_function (Sass_Function_Entry entry);
 ADDAPI void ADDCALL sass_delete_function_list (Sass_Function_List list);
 
 // Setters and getters for callbacks on function lists
-ADDAPI Sass_Function_Entry ADDCALL sass_function_get_list_entry(Sass_Function_List list, size_t pos);
-ADDAPI void ADDCALL sass_function_set_list_entry(Sass_Function_List list, size_t pos, Sass_Function_Entry cb);
+ADDAPI Sass_Function_Entry ADDCALL sass_function_get_list_entry(Sass_Function_List list, uint32_t pos);
+ADDAPI void ADDCALL sass_function_set_list_entry(Sass_Function_List list, uint32_t pos, Sass_Function_Entry cb);
 
 // Getters for custom function descriptors
 ADDAPI const char* ADDCALL sass_function_get_signature (Sass_Function_Entry cb);
